+++ lib/src/platform/io/niddler_server_announcement_manager.dart
@@ -7,10 +7,9 @@ import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
+import 'package:niddler_dart/src/platform/io/niddler_io.dart';
+import 'package:niddler_dart/src/platform/io/niddler_server.dart';
 import 'package:synchronized/synchronized.dart';
-
-import 'niddler.dart';
-import 'niddler_server.dart';
 
 const int _ANNOUNCEMENT_SOCKET_PORT = 6394;
 const int _COMMAND_REQUEST_QUERY = 0x01;
@@ -80,7 +79,7 @@ class NiddlerServerAnnouncementManager {
           streamer.add(1);
           await streamer.close();
         } finally {
-          await Future.delayed(Duration(seconds: 1));
+          await Future.delayed(const Duration(seconds: 1));
         }
       }
       await streamer.stream.first;
@@ -153,9 +152,8 @@ class NiddlerServerAnnouncementManager {
   static Future<void> _handleAnnounce(Stream<List<int>> socket, Future done,
       List<int> initialData, List<_Slave> slaves) async {
     final allDataBlobs = await socket.toList();
-    final allData = List<int>();
-    initialData.removeAt(0);
-    allData.addAll(initialData);
+    final allData = List<int>()
+      ..addAll(initialData.getRange(1, initialData.length));
     allDataBlobs.forEach(allData.addAll);
 
     final byteBuffer = Int8List.fromList(allData).buffer;
@@ -186,7 +184,7 @@ class NiddlerServerAnnouncementManager {
       }
     }
 
-    final slave = _Slave(socket, packageName, port, pid, protocolVersion, icon);
+    final slave = _Slave(packageName, port, pid, protocolVersion, icon);
     slaves.add(slave);
     // ignore: unawaited_futures
     done.then((_) => slaves.remove(slave));
@@ -246,13 +244,12 @@ class NiddlerServerAnnouncementManager {
 }
 
 class _Slave {
-  final Socket socket;
   final String packageName;
   final int port;
   final int pid;
   final int protocolVersion;
   final String icon;
 
-  _Slave(this.socket, this.packageName, this.port, this.pid,
-      this.protocolVersion, this.icon);
+  _Slave(
+      this.packageName, this.port, this.pid, this.protocolVersion, this.icon);
 }